verder gewerkt aan Movement.cpp, path van main.cpp gewijzigd

+++ codebase/Movement.h
@@ -9,9 +9,9 @@
 
 class Movement {
 private:
-    int8_t MOTOR_LEFT = PORT_B;
-    int8_t MOTOR_RIGHT = PORT_C;
-    int8_t currSpeed = 0;
+    uint8_t MOTOR_LEFT = PORT_B;
+    uint8_t MOTOR_RIGHT = PORT_C;
+    uint8_t currSpeed = 0;
 public:
     /**
      * Initializes Motors
@@ -34,7 +34,7 @@ public:
      * Sets the speed for the motors
      * @param speed (arbitrary) speed to set
      */
-    void speed(int32_t speed);
+    void speed(uint8_t speed);
 };
